refactored stepper and evaluator into separate modules
