+++ src/lib/exports.ts
@@ -1,3 +1,10 @@
+import * as L from '../lang.js'
+import { imageLib } from './image.js'
+import { musicLib } from './music.js'
+
 export { preludeEnv } from './prelude.js'
-export { imageLib } from './image.js'
-export { musicLib } from './music.js'
+
+export const internalLibs: Map<string, L.Env> = new Map([
+  ['image', imageLib],
+  ['music', musicLib]
+])
