+++ src/lib/music.ts
@@ -2,7 +2,8 @@
 import * as L from '../lang.js'
 import { msg } from '../messages.js'
 import { ok } from '../result.js'
-import { evaluateExp, runtimeError } from '../runtime.js'
+import { runtimeError } from '../runtime.js'
+import { evalExp } from '../evaluator.js'
 import * as Docs from './docs.js'
 import * as Pretty from '../pretty.js'
 import * as Utils from './utils.js'
@@ -37,7 +38,7 @@ const repeatPrim: L.Prim = async (env, args, app) =>
     if (n < 0) {
       return runtimeError(msg('error-precondition-not-met', 'repeat', 1, 'non-negative integer', Pretty.expToString(0, L.nlevalue(args[0]))))
     } else {
-      return evaluateExp(env,
+      return evalExp(env,
         L.nleif(
           L.nlecall(L.nlevar('='), [L.nlevalue(n), L.nlenumber(0)]),
           L.nlevar('empty'),
